requirements.txt
pillow
opencv-python
numpy
scikit-image
ultralytics
huggingface_hub
tqdm